requirements.txt
numpy>=1.26.0  # Umjesto numpy==1.24.
scikit-surprise==1.1.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
cython==0.29.36